refactor(recruitment): retire the Corporation/Recruitment namespace
The recruitment feature is a first-class domain (Pages/Recruitment,
Pages/Employment, Http/Controllers/Recruitment), but the review detail and its
shared inspection components were stranded under the defunct Corporation/Recruitment
namespace + /corporation/recruitment URL prefix. Move them home; delete the dead
enlistment-config island (superseded by Manage Recruitment / PostingController).

Backend
- ApplicationsController + ImpersonateRecruit → Http\Controllers\Recruitment.
- getApplication() renders 'Recruitment/Review/Application' (was Corporation/…).
- Surviving routes moved into the /recruitment group (routes/Routes/Recruitment.php),
  keeping every route NAME (all Vue/tests call route('name'), so URLs change with
  zero call-site edits). Collision-safe URLs: reviewer actions under
  /recruitment/application/{id}[/log|/update|/impersonate]; dashboard/onboarding
  apply+withdraw under the plural /recruitment/applications… (never clashes with the
  portal's POST /apply or DELETE /application/{id}). Old Routes/Corporation/Recruitment.php
  + its include removed.
- Deleted EnlistmentsController (open/close/save now live in PostingController).

Frontend
- Application.vue → Pages/Recruitment/Review/Application.vue.
- TabComponent + Tabs/* + UpdateCharacterComponent → Shared/Components/CharacterInspection/
  (shared by the review page AND Employment/Inspect observation; mirrors the backend
  CharacterInspectionScrollProps). Importers + the Wayfinder action import updated.
- Deleted the dead Configuration/* pages and the orphaned ActivityLog.vue / Applicant.vue.
- Pages/Corporation/Recruitment/ removed (Pages/Corporation keeps MemberTracking + Wallets).

Tests
- Repointed createEnlistment() + the two enlistment-CRUD cases to recruitment.posting.open/close;
  removed the redundant 'senior hr can setup watchlist' case (covered by ManageRecruitmentTest).
- Component assertions → 'Recruitment/Review/Application'; browser review URLs → /recruitment/application/{id}.
