+++ resources/js/Shared/Components/CharacterInspection/Tabs/AssetTab.vue
