+++ resources/js/Shared/Components/CharacterInspection/Tabs/ContractTab.vue
