+++ resources/js/Shared/Components/CharacterInspection/Tabs/LogTab.vue
