+++ resources/js/Shared/Components/CharacterInspection/Tabs/WalletTab.vue
