+++ routes/routes.php
@@ -82,7 +82,6 @@ Route::middleware('web')
                     ->group(function () {
                         include __DIR__.'/Routes/Corporation/Wallet.php';
                         include __DIR__.'/Routes/Corporation/MemberTracking.php';
-                        include __DIR__.'/Routes/Corporation/Recruitment.php';
                     });
 
                 Route::prefix('recruitment')
